Restructured MVC model, created pick up and move as well as GameFlow method
