+++ build.gradle
@@ -25,7 +25,7 @@ repositories {
 
 dependencies {
 
-    implementation 'org.json:json:20171018'
+    implementation 'org.json:json:20230227'
 
     implementation fileTree(dir: 'libs', include: '*.jar')
     implementation 'com.fasterxml.jackson.core:jackson-databind:2.14.2'
