+++ gradle.properties
@@ -5,9 +5,9 @@
 group=com.tlglearning.amnesiahospital
 version=1.0.0-SNAPSHOT
 
-# TODO: Replace "replace.me.project.AmnesiaHospitalGame" with the fully qualified class name of the project's main
+# TODO: Replace "replace.me.project.com.tlglearning.amnesiahospital.AmnesiaHospitalGame" with the fully qualified class name of the project's main
 #  class, without any quotes, as the value of the mainClass property.
-mainClass=AmnesiaHospitalGame
+mainClass=com.tlglearning.amnesiahospital.com.tlglearning.amnesiahospital.AmnesiaHospitalGame
 
 javaVersion=11
 shadowVersion=8.1.0
