--- src/main/java/Weapon.java
@@ -1,5 +1,0 @@
-public class Weapon {
-
-  
-
-}
