+++ src/main/java/com/tlglearning/amnesiahospital/model/Item.java
@@ -1,3 +1,5 @@
+package com.tlglearning.amnesiahospital.model;
+
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
