+++ src/main/java/com/tlglearning/amnesiahospital/model/ItemType.java
@@ -1,3 +1,5 @@
+package com.tlglearning.amnesiahospital.model;
+
 public enum ItemType {
   HEALTH_SERUM,
