+++ src/main/java/com/tlglearning/amnesiahospital/model/Weapon.java
@@ -1,0 +1,7 @@
+package com.tlglearning.amnesiahospital.model;
+
+public class Weapon {
+
+  
+
+}
